[logiclayer] removes shared ll file
