--- tesseract-server/src/handlers/logic_layer/shared.rs
@@ -1,103 +1,0 @@
-use actix_web::{
-    FutureResponse,
-    HttpResponse,
-};
-use futures::future::{self};
-use failure::{Error, format_err};
-
-
-#[derive(Debug, Clone)]
-pub enum TimeValue {
-    First,
-    Last,
-    Value(u32),
-}
-
-impl TimeValue {
-    pub fn from_str(raw: String) -> Result<Self, Error> {
-        if raw == "latest" {
-            Ok(TimeValue::Last)
-        } else if raw == "oldest" {
-            Ok(TimeValue::First)
-        } else {
-            match raw.parse::<u32>() {
-                Ok(n) => Ok(TimeValue::Value(n)),
-                Err(_) => Err(format_err!("Wrong type for time argument."))
-            }
-        }
-    }
-}
-
-
-#[derive(Debug, Clone)]
-pub enum TimePrecision {
-    Year,
-    Quarter,
-    Month,
-    Week,
-    Day,
-}
-
-impl TimePrecision {
-    pub fn from_str(raw: String) -> Result<Self, Error> {
-        match raw.as_ref() {
-            "year" => Ok(TimePrecision::Year),
-            "quarter" => Ok(TimePrecision::Quarter),
-            "month" => Ok(TimePrecision::Month),
-            "week" => Ok(TimePrecision::Week),
-            "day" => Ok(TimePrecision::Day),
-            _ => Err(format_err!("Wrong type for time precision argument."))
-        }
-    }
-}
-
-
-#[derive(Debug, Clone)]
-pub struct Time {
-    pub precision: TimePrecision,
-    pub value: TimeValue,
-}
-
-impl Time {
-    pub fn from_str(raw: String) -> Result<Self, Error> {
-        let e: Vec<&str> = raw.split(".").collect();
-
-        if e.len() != 2 {
-            return Err(format_err!("Wrong format for time argument."));
-        }
-
-        let precision = match TimePrecision::from_str(e[0].to_string()) {
-            Ok(precision) => precision,
-            Err(err) => return Err(err),
-        };
-        let value = match TimeValue::from_str(e[1].to_string()) {
-            Ok(value) => value,
-            Err(err) => return Err(err),
-        };
-
-        Ok(Time {precision, value})
-    }
-
-    pub fn from_key_value(key: String, value: String) -> Result<Self, Error> {
-        let precision = match TimePrecision::from_str( key) {
-            Ok(precision) => precision,
-            Err(err) => return Err(err),
-        };
-        let value = match TimeValue::from_str(value) {
-            Ok(value) => value,
-            Err(err) => return Err(err),
-        };
-
-        Ok(Time {precision, value})
-    }
-}
-
-
-/// Helper method to return errors (FutureResponse<HttpResponse>).
-pub fn boxed_error(message: String) -> FutureResponse<HttpResponse> {
-    Box::new(
-        future::result(
-            Ok(HttpResponse::NotFound().json(message))
-        )
-    )
-}
