+++ tesseract-server/src/logic_layer/mod.rs
@@ -2,6 +2,6 @@ mod cache;
 mod config;
 mod util;
 
-pub use self::cache::{Cache, CubeCache, populate_cache};
+pub use self::cache::{Cache, CubeCache, Time, TimePrecision, TimeValue, populate_cache};
 pub use self::config::{LogicLayerConfig, read_config};
-pub use self::util::{format_column_data};
+pub use self::util::{stringify_column_data, boxed_error};
